requirements.txt
discord.py==1.7.3
flask
async-timeout==3.0.1